updated edit and add meal
updated the structure of files
