+++ client/src/components/pages/Kitchen Admin/Add Meal/addMeal.jsx
@@ -1,0 +1,14 @@
+import React from "react";
+import NavBar from "../../../organisms/Kitchen/header/header";
+import Content from "../../../organisms/Kitchen/MealDetails/MealDetailsForm";
+
+const MealAdd = () => {
+  return (
+    <div>
+      <NavBar />
+      <Content />
+    </div>
+  );
+};
+
+export default MealAdd;
