+++ client/src/components/pages/Kitchen Admin/Edit Meal/editMeal.jsx
@@ -1,0 +1,19 @@
+import React from "react";
+
+import NavBar from "../../../organisms/Kitchen/header/header";
+
+import ContentEdit from "../../../organisms/Kitchen/EditMeal/editMeal";
+
+const MealEdit = () => {
+  return (
+    <div>
+         <NavBar/>
+         <ContentEdit/>
+       
+    </div>
+
+
+  );
+};
+
+export default MealEdit;
